+++ NEWS.md
@@ -8,6 +8,7 @@
 * Rewrote kde_bandwidth() to handle more methods
 * Added gg_density() and gg_density_layer()
 * Replaced density_scores() with surprisals()
+* Updated gg_bagplot() and gg_hdrboxplot() to use show_points argument
 * Removed lookout()
 * Added hampel_anomalies()
 * mvscale() moved to lookout package